UPDATE: FUSION DE CLASE PRODUCTOPRESENTACION Se cambio la antigua clase productopresentacionDAO por la mejorada de la aplicacion DESKTOP y se esta implementando la funcionalidad de stock en cada venta que se realize, algunas funciones del punto de venta se vieron afectados, se debe depurar las vistas de ventas en las distintas cajas.

+++ src/Modelo/VentaProductoDAO.java
@@ -271,7 +271,7 @@ public class VentaProductoDAO extends Conexion implements VentaProductoCRUD {
         }
     }
 
-    public boolean sumarStock(int idProdPresentacion, int cant, int tipoVenta) throws Exception {
+    public boolean sumarStock(int idProdPresentacion, double cant, int tipoVenta) throws Exception {
         //tipoVenta ==> 1= venta Real; 2= venta nota pedido
         String stock = "stock";
         if (tipoVenta == 2) {
